[get_free-energy_extrema.py]: Change Prominence for Peak Finding
Choose the prominence such that 100*`p_min` percent of the particles
have a higher kinetic energy in the z direction than the prominence.
Thus, `p_min` is now the adjustable parameter and `prominence` should
not be touched anymore.  Setting `p_min` to 0.5 results in a prominence
of 0.2275 which is close to the previous value of 0.25 but now as a
better physical reason.

Also see the notes on pages 142-147 in my "derivation book".