[get_free-energy_extrema.py]: Ensure Layering Region Ends With Maximum
Ensure that the layering region and bulk region are separated by a
free-energy maximum.  The layering region must not end with a minimum,
because then there exists a free-energy barrier that a particle has to
overcome when traveling from the electrode to the bulk that is higher
than the prominence threshold.