[plot_bin_lifetimes_chain-length_dependence_cluster.py]: Create
Create a new Python script that plots the bin residence times of a given
compound as function of the PEO chain length.

Only bins that correspond to actual layers (i.e. free-energy minima) at
the electrode interface are taken into account.

Layers are clustered based on their distance to the electrode.