[plot_bin_lifetimes_surfq_dependence_cluster.py]: Create
Create a Python script that plots the bin residence times of a given
compound as function of the electrode surface charge

Only bins that correspond to actual layers (i.e. free-energy minima) at
the electrode interface are taken into account.

Layers/free-energy minima are clustered based on their distance to the
electrode.