[io_handler.py]: Update Functions `peek_edr` and `read_edr`
Update the functions `lintf2_ether_ana_postproc.io_handler.peek_edr` and
`lintf2_ether_ana_postproc.io_handler.read_edr`:

Remove "/mol" from units, because I did not specify the number of
molecules with the `-nmol` option when calling `gmx energy`.  Thus, the
energies in the .edr files are the total energies of the system and not
energies per molecule.